Refactor NaN detection and debugging utilities for improved clarity and functionality
